create a validationUtils file for validation utilities.
