+++ validations/validationUtils.js
@@ -1,0 +1,15 @@
+import { validationResult } from "express-validator";
+
+function validateRequest(req, res, next) {
+  const errors = validationResult(req);
+  if (errors.isEmpty()) {
+    return next();
+  }
+  return res.status(400).json({ errors: errors.array() });
+}
+
+function withValidation(validations) {
+  return [...validations, validateRequest];
+}
+
+export { withValidation };
